refactor: rename for initial release
