+++ curator/data_pipeline/polygon_options.json
